test(oss-licenses): Isolate V2 tests to avoid static state pollution
